+++ types/adhoc.pb.go
@@ -53,6 +53,8 @@
 		HookList
 		KeepaliveRecord
 		Metrics
+		MetricPoint
+		MetricTag
 		Mutator
 		Organization
 		Rule