+++ CHANGELOG.md
@@ -18,6 +18,7 @@ set when importing legacy settings.
 - Add CPU architecture in system information of entities.
 - The `sensuctl user change-password` subcommand now accepts flag parameters.
 - Configured and enabled etcd autocompaction.
+- Add event metrics type, implementing the Sensu Metrics Format.
 
 ### Changed
 - Refactor Check data structure to not depend on CheckConfig. This is a breaking